+++ ChangeLog.md
@@ -113,6 +113,7 @@
 - BUGFIX: #5609 Fixed SQL syntax highlighter for multi-line comments with empty linies inside.
 - BUGFIX: #5628 Fixed PDF export bugs (first page size too small, occasional table overlapping, page numbers cut off).
 - BUGFIX: #5691 Fixed tooltip in Database Dialog with Wayland under Linux.
+- BUGFIX: #5692 Select first inserted row after inserting multiple rows in Grid View.
 
 ### 3.4.21
 - ADDED: #5466 Support foreign keys with implicit columns (proper values in FK combobox columns).